requirements.txt
baidu-aip==2.2.13.0